refactor: getSPLToken
